avoid working on non-existing dir

+++ public_html/lists/admin/lib.php
@@ -1415,6 +1415,7 @@ function delFsTree($dir) {
   if (!is_dir($dir) && is_file($dir)) {
     return unlink($dir); ## delete file
   }
+  if (!is_dir($dir)) return;
   $files = array_diff(scandir($dir), array('.','..'));
   foreach ($files as $file) {
     (is_dir("$dir/$file")) ? delFsTree("$dir/$file") : unlink("$dir/$file");
